// types/serverStatus.ts
export type ServerStatus = {
    passworded: boolean;
    serverName: string;
    forkName: string;
    buildVersion: number;
    currentMap: string;
    gameMode: string;
    ingameTime: string;
    roundTime: string;
    playerCount: number;
    playerCountMax: number;
    serverIP: string;
    serverPort: number;
    winDownload: string;
    osxDownload: string;
    linuxDownload: string;
    fps: number;
    goodFileVersion: string;
};